Represent int4 as the LLVM type i4 instead of i8.
Before, int4 was represented as i8. But this caused issues when converting S8 to S4 then back to S8 within a single fusion, as the S4 value was represented as i8 and so could hold values not representable with a signed int4 value. This could have been solved by introducing a mask when converting to S4, but I think it's cleaner to directly use i4, as this avoids the unusual case where the XLA type did not match the corresponding LLVM type.

PiperOrigin-RevId: 583238198